move statistic related to different file
